hr-chatbot/3-fixed-chatbot/requirements.txt
arize-phoenix
arize-phoenix-otel
openai
openinference-instrumentation-openai
python-dotenv